// CreateNow/my_app/src/config/types/event.ts
export interface NewEventType {
    
    autor?: string;
    id?: string;
    catering?: string;
    miasto?: string;
    dodatki?: string;
    fotografia?: string;
    kamerzysta?: string;
    lokale?: string;
    muzyka?: string;
    samochody?: string;
    
  }
  
  export interface AddEventType {
    
    autor: string;
    catering: string;
    miasto: string;
    dodatki: string;
    fotografia: string;
    kamerzysta: string;
    lokale: string;
    muzyka: string;
    samochody: string;
    
  }